+++ test_apps/weight_tracker/lib/main.app.framy.dart
@@ -1094,6 +1094,7 @@ class _FramyCustomPageState extends State<FramyCustomPage> {
                 child: DevicePreview(
                   enabled: FramyAppSettings.of(context).wrapWithDevicePreview,
                   style: DevicePreviewStyle(
+                    hasFrameShadow: false,
                     toolBar: DevicePreviewToolBarStyle.light(
                       position: DevicePreviewToolBarPosition.right,
                     ),
@@ -1512,13 +1513,24 @@ class FramyStoryboardCustomPageWithDependencies extends StatelessWidget {
     return Column(
       children: [
         Expanded(
-          child: Container(
-            decoration: BoxDecoration(
-              border: Border.all(color: Colors.black54),
-              borderRadius: BorderRadius.circular(16),
-            ),
-            child: builder(dependencyValue),
-          ),
+          child: FramyAppSettings.of(context).wrapWithDevicePreview
+              ? DevicePreview(
+                  style: DevicePreviewStyle(
+                    hasFrameShadow: false,
+                    toolBar: DevicePreviewToolBarStyle.light(),
+                    background: BoxDecoration(),
+                  ),
+                  isToolBarVisible: false,
+                  areSettingsEnabled: false,
+                  builder: (_) => builder(dependencyValue),
+                )
+              : Container(
+                  decoration: BoxDecoration(
+                    border: Border.all(color: Colors.black54),
+                    borderRadius: BorderRadius.circular(16),
+                  ),
+                  child: builder(dependencyValue),
+                ),
         ),
         const SizedBox(height: 8),
         Text(name, style: Theme.of(context).textTheme.caption),